server/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pytest
httpx
pip freeze > requirements.txt
